requirements.txt
opencv-python
numpy
mediapipe
pynput
PyGLM